+++ database/migrations/2021_06_10_135553_create_b_i_d_compares_table.php
@@ -18,8 +18,8 @@ class CreateBIDComparesTable extends Migration
             $table->unsignedBigInteger('order_id')->index();
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
             $table->bigInteger('amount');
-            $table->float('start_rate');
-            $table->float('end_rate');
+            $table->double('start_rate');
+            $table->double('end_rate');
             $table->bigInteger('status');
             $table->timestamps();
         });
